+++ node/src/lib.rs
@@ -23,6 +23,7 @@ extern crate alloc;
 mod aura_or_nimbus_consensus;
 pub mod chain_specs;
 pub mod cli;
+pub mod client;
 pub mod command;
 pub mod rpc;
 pub mod service;